remove old console tab code
