+++ pgmanage/app/static/assets/js/pgmanage_frontend/src/console.js
@@ -293,13 +293,7 @@ function consoleHistoryLastPage() {
 function consoleHistoryOpenCmd(p_index) {
 	let command = v_connTabControl.selectedTab.tag.tabControl.selectedTab.tag.consoleHistory.grid.getDataAtRow(p_index)[1];
 	let selected_tab = v_connTabControl.selectedTab.tag.tabControl.selectedTab
-	if (selected_tab.tag.vueComponent) {
-		emitter.emit(`${selected_tab.id}_copy_to_editor`, command)
-	} else {
-		v_connTabControl.selectedTab.tag.tabControl.selectedTab.tag.editor_input.setValue(command);
-		v_connTabControl.selectedTab.tag.tabControl.selectedTab.tag.editor_input.clearSelection();
-		v_connTabControl.selectedTab.tag.tabControl.selectedTab.tag.editor_input.gotoLine(0, 0, true);
-	}
+	emitter.emit(`${selected_tab.id}_copy_to_editor`, command)
 	closeConsoleHistory();
 }
