+++ pgmanage/app/static/assets/js/pgmanage_frontend/src/create_tab_functions.js
@@ -28,7 +28,6 @@ SOFTWARE.
 import { v_createSnippetTextTabFunction } from './tab_functions/inner_snippet_tab'
 import { v_createConnTabFunction } from './tab_functions/outer_connection_tab'
 import { v_createQueryTabFunction } from './tab_functions/inner_query_tab'
-import { v_createConsoleTabFunction } from './tab_functions/inner_console_tab'
 import { v_createSnippetPanelFunction } from './panel_functions/outer_snippet_panel'
 import { createEditDataTabFunction } from './tab_functions/inner_edit_data_tab'
 import { v_createMonitoringTabFunction } from './tab_functions/inner_monitoring_tab'
@@ -40,7 +39,7 @@ import { createERDTabFunction } from './tab_functions/erd_tab'
 import ContextMenu from '@imengyu/vue3-context-menu'
 import { createDebuggerTabFunction } from './tab_functions/inner_debugger_tab'
 import { showConfirm } from './notification_control'
-import { createConsoleTabFunctionNew } from './tab_functions/inner_console_tab_new'
+import { createConsoleTabFunction } from './tab_functions/inner_console_tab'
 
 function initCreateTabFunctions() {
 
@@ -422,8 +421,7 @@ function initCreateTabFunctions() {
 
   // Functions to create tabs inside a connection tab
 	v_connTabControl.tag.createQueryTab = v_createQueryTabFunction;
-  v_connTabControl.tag.createConsoleTabNew = createConsoleTabFunctionNew;
-	v_connTabControl.tag.createConsoleTab = v_createConsoleTabFunction;
+  v_connTabControl.tag.createConsoleTab = createConsoleTabFunction;
 
 //   v_connTabControl.tag.createWebsiteTab = v_createWebsiteTabFunction;
   v_connTabControl.tag.createNewMonitorUnitTab = v_createNewMonitorUnitTabFunction;
